Add Dockerfile for independent parflow image
This parflow image is intended to be ran as a separate image, rather
than being a part of our cluster image.

The runtime is created independently from the build image to
minimize the size of the final docker image, which is currently only
158 MB.

Some tests can be ran in the parflow repo's test directory, and they seem
to work. For instance:
```
docker run --rm -w /data -v $PWD:/data kitware/hpccloud:parflow default_single.tcl 1 1 1
```

This Dockerfile can be used to create other types of images as well, such
as singularity and shifter.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>