Add singularity to our slurm cluster
This adds singularity to our slurm cluster, with which you can build/pull
singularity images, and then run them.

Here is an example of using it:

    # Start up the cluster:
    docker-compose up

    # In a separate terminal copy the parflow test directory to the compute-pyfr docker container
    docker cp test <id>:/home/demo/

    # Enter into the compute-pyfr container
    docker exec -u demo -w /home/demo/test -it <id> bash

    # Build the singularity image. The docker image must be present locally on your computer.
    # This image comes from https://github.com/Kitware/hpccloud-services/pull/14
    singularity build parflow.sif docker-daemon:kitware/hpccloud:parflow

    # Run a test
    singularity run ./parflow.sif default_overland.tcl 1 1 1
    # default_overland : PASSED

Note: the cluster images must be ran in "privileged" mode in order
to use singularity on them. That is being done in the docker-compose files
now.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>